Use Tox for running the tests and linting in CI
This makes the tests pass in CI because for some Python versions
SQLAlchemy 1.4 gets installed which SQLAlchemy-Searchable does not
support. This also makes the CI environment more consistent with how the
tests are run locally and it will make it simple add SQLAlchemy 1.4 to
the CI test matrix once SQLAlchemy-Seachable supports it.